+++ app/src/main/java/com/babylon/wallet/android/domain/usecases/signing/ROLAClient.kt
@@ -2,6 +2,7 @@ package com.babylon.wallet.android.domain.usecases.signing
 
 import com.babylon.wallet.android.domain.model.signing.SignPurpose
 import com.babylon.wallet.android.domain.model.signing.SignRequest
+import com.babylon.wallet.android.domain.model.transaction.UnvalidatedManifestData
 import com.babylon.wallet.android.domain.usecases.assets.GetEntitiesOwnerKeysUseCase
 import com.babylon.wallet.android.domain.usecases.transaction.GenerateAuthSigningFactorInstanceUseCase
 import com.babylon.wallet.android.presentation.accessfactorsources.AccessFactorSourcesInput
@@ -15,7 +16,6 @@ import com.radixdlt.sargon.extensions.ProfileEntity
 import com.radixdlt.sargon.extensions.hex
 import com.radixdlt.sargon.extensions.init
 import com.radixdlt.sargon.extensions.setOwnerKeysHashes
-import rdx.works.core.domain.UnvalidatedManifestData
 import rdx.works.core.sargon.transactionSigningFactorInstance
 import javax.inject.Inject
