+++ app/src/main/java/com/babylon/wallet/android/domain/usecases/transaction/SendClaimRequestUseCase.kt
@@ -1,7 +1,8 @@
 package com.babylon.wallet.android.domain.usecases.transaction
 
 import com.babylon.wallet.android.data.dapp.IncomingRequestRepository
-import com.babylon.wallet.android.data.manifest.prepareInternalTransactionRequest
+import com.babylon.wallet.android.domain.model.transaction.UnvalidatedManifestData
+import com.babylon.wallet.android.domain.model.transaction.prepareInternalTransactionRequest
 import com.radixdlt.sargon.Account
 import com.radixdlt.sargon.NonFungibleResourceAddress
 import com.radixdlt.sargon.TransactionManifest
@@ -10,7 +11,6 @@ import com.radixdlt.sargon.extensions.orZero
 import com.radixdlt.sargon.extensions.stakesClaim
 import com.radixdlt.sargon.extensions.string
 import com.radixdlt.sargon.extensions.sumOf
-import rdx.works.core.domain.UnvalidatedManifestData
 import rdx.works.core.domain.assets.StakeClaim
 import rdx.works.core.domain.resources.Resource
 import javax.inject.Inject
